Snackbar for errors, some refactoring
